Split keybind definitions into a separate file
